Discussion for user and Discussion_admin panel
The Discussion is working whenever a user post's a discussion. it will show them a message on the screen post submitted awaiting approval and from the admin side, the admin can approve or dis-approve the post of any user and whever the admin approve the post it will show that post on the discussion page and the data for both approve and disapprove will be submitted in the database. the admin can also delete and caontrol the post content from the database, moreover it will react on the discussion page.